+++ frontend/src/screens/index.js
@@ -1,5 +1,5 @@
 export { default as Products } from './Product';
-export { default as Cart } from './Cart';
+export { default as Cart } from './Cart/index';
 export { default as Landing } from './Landing';
 export { default as SingleProduct } from './SingleProduct';
-export { default as WishList } from './WishList';
+export { default as WishList } from './Wishlist';
